+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 
 ## [0.15.1] - 2025-02-04
 ### Added    
-* `cases` kwarg to `addVarStat` to allow specifying which cases to use to calculate the stat ([GH-14](https://github.com/scottshambaugh/monaco/issues/14))
+* `cases` kwarg to `addVarStat` and `calcSensitivities` to allow specifying which cases to use to calculate the stat ([GH-14](https://github.com/scottshambaugh/monaco/issues/14))
 
 ## [0.15.0] - 2025-02-03
 ### Added