+++ CHANGELOG.md
@@ -25,9 +25,9 @@
 ## [Unreleased]
 ### Added    
 * `case.vars` dict to reference all invars and outvars, and `case.vals` dict to reference all invals and outvals
-* `__getitem__` for cases to get a val by name
-* `__getitem__` for sims to get a case by number or a var by name
-* `__getitem__` for vars to get a val by number
+* `__getitem__` for cases to get a val by name, e.g. `case['valname']`
+* `__getitem__` for sims to get a case by number or a var by name, e.g. `sim[i]` or `sim['varname']`
+* `__getitem__` for vars to get a val by number, e.g. `var[i]`
 * `datasource` attribute on invals and outvals to tell if they were loaded from file
 ### Changed    
 ### Removed